t6

+++ hello_world/6-size.c
@@ -12,10 +12,10 @@ int main(void)
 	long long int d;
 	float e;
 
-	printf("Size of a char: %zu byte(s)\n", (unsigned long)sizeof(a));
-	printf("Size of a int: %zu byte(s)\n", (unsigned long)sizeof(b));
-	printf("Size of a long int: %zu byte(s)\n", (unsigned long)sizeof(c));
-	printf("Size of a long long int: %zu byte(s)\n", (unsigned long)sizeof(d));
-	printf("Size of a float: %zu byte(s)\n", (unsigned long)sizeof(e));
+	printf("Size of a char: %zu byte(s)\n", sizeof(a));
+	printf("Size of a int: %zu byte(s)\n", sizeof(b));
+	printf("Size of a long int: %zu byte(s)\n", sizeof(c));
+	printf("Size of a long long int: %zu byte(s)\n", sizeof(d));
+	printf("Size of a float: %zu byte(s)\n", sizeof(e));
 	return (0);
 }
